requirements.txt
flask==3.0.3
requests==2.31.0
pyyaml==6.0.1
jinja2==3.1.2
colorama==0.4.6
python-dotenv==1.0.0
